+++ Assets/Scripts/UI/GarageMenu.cs
@@ -24,15 +24,13 @@ public class GarageMenu : MonoBehaviour
     [SerializeField] private TextMeshProUGUI engineLevelText, handlingLevelText, nosLevelText;
     // Menu buttons
     [SerializeField] private GameObject changeCarButton, selectItemButton, purchaseButton;
-    // NFT sprites
-    [SerializeField] private Texture2D Nft1, Nft2, Nft3;
 
     #endregion
 
     #region Methods
     
     /// <summary>
-    /// Initializes objects and sets our default colour
+    /// Initializes objects
     /// </summary>
     private void Awake()
     {
@@ -56,7 +54,7 @@ public class GarageMenu : MonoBehaviour
     }
     
     /// <summary>
-    /// A function to plat the menu selection sound
+    /// A function to play the menu selection sound
     /// </summary>
     public void PlayMenuSelect()
     {
@@ -150,36 +148,6 @@ public class GarageMenu : MonoBehaviour
         PlayMenuSelect();
     }
 
-    /// <summary>
-    /// Changes to NFT 1
-    /// </summary>
-    public void SelectNft1()
-    {
-        globalManager.nftSprite = Nft1;
-        // Play our menu select audio
-        PlayMenuSelect();
-    }
-
-    /// <summary>
-    /// Changes to NFT 2
-    /// </summary>
-    public void SelectNft2()
-    {
-        globalManager.nftSprite = Nft2;
-        // Play our menu select audio
-        PlayMenuSelect();
-    }
-
-    /// <summary>
-    /// Changes to NFT 3
-    /// </summary>
-    public void SelectNft3()
-    {
-        globalManager.nftSprite = Nft3;
-        // Play our menu select audio
-        PlayMenuSelect();
-    }
-    
     /// <summary>
     /// Purchase manager for upgrades
     /// </summary>
