+++ Assets/Scripts/UI/MarketplaceMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 720b694365af57e4d89b3c549b301e48
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
