+++ Assets/Scripts/UI/NftMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6ffdb236a9c5626409bada4487cea3e3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
